update: 認証API関連の関数を一つにまとめる
