--- front/src/lib/api/auth/getCurrentUser.ts
@@ -1,19 +1,0 @@
-import Cookies from "js-cookie";
-
-import { client } from "lib/api/client";
-
-//認証済みのユーザーを取得
-export const getCurrentUser = () => {
-  if (
-    !Cookies.get("_access_token") ||
-    !Cookies.get("_client") ||
-    !Cookies.get("_uid")
-  )
-    return client.get("/auth/sessions", {
-      headers: {
-        "access-token": Cookies.get("_access_token"),
-        client: Cookies.get("_client"),
-        uid: Cookies.get("_uid"),
-      },
-    });
-};
