--- front/src/lib/api/auth/signOut.ts
@@ -1,14 +1,0 @@
-import Cookies from "js-cookie";
-
-import { client } from "lib/api/client";
-
-//サインアウト（ログアウト）
-export const signOut = () => {
-  return client.delete("/auth/sign_out", {
-    headers: {
-      "access-token": Cookies.get("_access_token"),
-      client: Cookies.get("_client"),
-      uid: Cookies.get("_uid"),
-    },
-  });
-};
